refactor(persistence): moved MongoTaskDaoAdapter to 'infrastructure.mongodb'
moved TaskDao to 'common.persistence'

 removed obsolete DummyMongo class
