--- src/main/java/infrastructure/mongo/dao/DummyMongo.java
@@ -1,4 +1,0 @@
-package infrastructure.mongo.dao;
-
-public class DummyMongo {
-}
